+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - The protoc-commands command is now accessible via the `dry-run`
   flag for compile, format, gen, and lint.
+- The list-* commands are now accessible via lint command flags:
+  list-all, list-configured, list-groups, and list-group.
 
 ## [0.4.0] - 2018-06-22
 ### Added